Only list id, type and name for part-relations
I would have found it nice to recursively refer to the main schema
(schema.json) and then change the required properties as we would only
require an ID here. As you can not overwrite the "required" config from a
referenced schema, I copied over the main schema, see
14f4311bc7788d01681396ca1aa263741058b259.
However, this led to all the fields being displayed in the SkoHub
Editor. With this commit, I switch to only list id, type and name in the
isPartOf/hasPart schemas.